Updated title to show remaining time and cleaned up existing code.
Cleaned up the code a little.  I made the minutesRemaining and secondsRemaining a global variable that can be used throughout the code instead of relying on reading the values from the timer and parsing it out.  I'm populating the title of the document now so you can see the remaining time in the tab while you're working on other things.  I'm added ability to update timer whenever incrementing or decrementing session/break values as long as timer is not currently running and hasn't started yet.